+++ saleor/checkout/webhooks/tests/test_shipping_list_methods_for_checkout.py
@@ -1,14 +1,11 @@
-import json
 from datetime import timedelta
 from decimal import Decimal
 from unittest import mock
 
-import graphene
 from django.utils import timezone
 
 from ....webhook.event_types import WebhookEventSyncType
 from ....webhook.payloads import generate_checkout_payload
-from ....webhook.transport.asynchronous import create_deliveries_for_subscriptions
 from ....webhook.transport.shipping import (
     get_cache_data_for_shipping_list_methods_for_checkout,
 )
@@ -234,39 +231,3 @@ def test_ignore_selected_fields_on_generating_cache_key(
         mocked_webhook_response,
         timeout=CACHE_TIME_SHIPPING_LIST_METHODS_FOR_CHECKOUT,
     )
-
-
-def test_shipping_list_methods_for_checkout(
-    checkout_with_shipping_required,
-    subscription_webhook,
-    address,
-    shipping_method,
-):
-    # given
-    query = """
-    subscription {
-      event {
-        ...on ShippingListMethodsForCheckout {
-          checkout {
-            id
-          }
-        }
-      }
-    }
-    """
-    event_type = ShippingListMethodsForCheckout.event_type
-    webhook = subscription_webhook(query, event_type)
-    checkout = checkout_with_shipping_required
-    checkout.shipping_address = address
-    checkout.shipping_method = shipping_method
-    webhooks = [webhook]
-    checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
-
-    # when
-    deliveries = create_deliveries_for_subscriptions(event_type, checkout, webhooks)
-
-    # then
-    payload = json.loads(deliveries[0].payload.get_payload())
-    assert payload["checkout"] == {"id": checkout_id}
-    assert len(deliveries) == len(webhooks)
-    assert deliveries[0].webhook == webhooks[0]
